# openapi.yaml
openapi: "3.0.3"
info:
  title: YumYum API
  description: "This API allows Yam-Yam users to make all the CRUD operations on the following resources: (Restaurants, Orders, Reviews, Users, Carts)."
  version: 1.0.0
servers:
  - url: "http://localhost:3000/api/v1"
...
paths:
...
  /users/{id}:
    parameters:
      - name: id
        in: path
        schema:
          type: string
        required: true
        example: "ObjectId('...')"

    get:
      tags:
        - Users
      summary: Read a User
      description: Use this endpoint to get a specific user by ID.
      operationId: getUser
      responses:
        "200":
          $ref: "#/components/responses/UserResponse"
        "4XX":
          $ref: "#/components/responses/ClientErrorResponse"
        "5XX":
          $ref: "#/components/responses/InternalServerErrorResponse"
        default:
          $ref: "#/components/responses/UnexpectedErrorResponse"
    put:
      tags:
        - Users
      summary: Update a User
      description: Use this endpoint to update an existing user in the database.
      operationId: updateUser
      requestBody:
        $ref: "#/components/requestBodies/UserRequest"
      responses:
        "200":
          $ref: "#/components/responses/UserResponse"
        "4XX":
          $ref: "#/components/responses/ClientErrorResponse"
        "5XX":
          $ref: "#/components/responses/InternalServerErrorResponse"
        default:
          $ref: "#/components/responses/UnexpectedErrorResponse"
    delete:
      tags:
        - Users
      summary: Delete a User
      description: Use this endpoint to delete an existing user from the database.
      operationId: deleteUser
      responses:
        "200":
          $ref: "#/components/responses/Success"
        "4XX":
          $ref: "#/components/responses/ClientErrorResponse"
        "5XX":
          $ref: "#/components/responses/InternalServerErrorResponse"
        default:
          $ref: "#/components/responses/UnexpectedErrorResponse"
components:
  requestBodies:
    UserRequest:
      description: User Details
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/User"
...
  responses:
    UserResponse:
      description: User Details
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/User"
...
    Success:
      description: Success
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Success"
    ClientErrorResponse:
      description: Client error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    UnexpectedErrorResponse:
      description: Unexpected error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    InternalServerErrorResponse:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    User:
      type: object
      properties:
        _id:
          type: string
          description: "User unique identifier"
          example: "ObjectId('...')"
        name:
          type: string
          description: "Name of the user"
          example: "John Doe"
        email:
          type: string
          description: "Email address of the user"
          example: "john@example.com"
        phone:
          type: string
          description: "Phone number of the user"
          example: "+201001234567"
        password:
          type: string
          description: "Hashed password of the user"
          example: "hashed_password"
        addresses:
          type: array
          items:
            $ref: "#/components/schemas/address"
        cart:
          type: string
          description: "Cart ObjectId, referring to the user's current shopping cart"
          example: "ObjectId('...')"
        orderHistory:
          type: array
          items:
            type: object
            properties:
              orderId:
                type: string
                description: "Order unique identifier"
                example: "ObjectId('...')"
        payments:
          type: array
          items:
            type: object
            properties:
              cardNumber:
                type: string
                description: "Encrypted credit card number"
                example: "encrypted_string"
              cvv:
                type: string
                description: "Encrypted CVV code"
                example: "encrypted_string"
              date:
                type: string
                description: "Expiration date of the card in MM/YY format"
                example: "12/24"
        imgUrl:
          type: string
          description: "URL to the user's profile image"
          example: "https://example.com/profile.jpg"
...
    address:
      type: object
      properties:
        street:
          type: string
          description: "Street address"
          example: "123 Main St."
        city:
          type: string
          description: "City"
          example: "Foodtown"
        addressInfo:
          type: string
          description: "Additional address information"
          example: "Building 1, Floor 2, Apt 3"
        buildingNo:
          type: string
          description: "Building number"
          example: "1"
        floorNo:
          type: string
          description: "Floor number"
          example: "2"
    Success:
      type: object
      properties:
        status:
          type: string
          example: success
    Error:
      type: object
      properties:
        status:
          type: string
          example: error
        message:
          type: string
          example: "Error details"
...